+++ src/api-sock.js
@@ -26,88 +26,33 @@
 
 \******************************************************************************/
 
-import Sock from './sock.js'
+import Sock       from './sock.js'
+import Subscriber from './subscriber.js'
 
-class Team {
+
+class TeamSubscriber extends Subscriber {
   constructor(sock, team, max_count = 10000) {
-    this.sock        = sock
-    this.team        = team
-    this.subscribers = {}
-    this.data        = []
-    this.max_count   = max_count
+    super(sock, max_count)
+    this.ref = `team-${team}`
+    this.msg = {timeseries: 'team.score', team, '$ref': this.ref}
   }
+}
 
 
-  get has_subscribers() {return 0 < Object.keys(this.subscribers).length}
-
-
-  add_subscriber(cb) {
-    let id = this.sock.nextID++
-    this.subscribers[id] = cb
-    this.data.map(cb)
-    this.update()
-    return id
+class UserSubscriber extends Subscriber {
+  constructor(sock, uid, pid, max_count = 10000) {
+    super(sock, max_count)
+    this.ref = `user-${uid}-${pid}`
+    this.msg = {timeseries: 'user.score', uid, pid, '$ref': this.ref}
   }
+}
 
 
-  del_subscriber(id) {
-    delete this.subscribers[id]
-    this.update()
+function get_chart_subscriber(sock, chart) {
+  switch (chart.type) {
+  case 'team': return new TeamSubscriber(sock, chart.team)
+  case 'user': return new UserSubscriber(sock, chart.uid, chart.pid)
   }
-
-
-  _get_message(action) {
-    return {
-      type:       action,
-      timeseries: 'team.score',
-      team:       this.team,
-      '$ref':     'team-' + this.team
-    }
-  }
-
-
-  _subscribe() {
-    if (!this.sock.connected) return this.sock.connect()
-
-    let msg = this._get_message('subscribe')
-    msg.max_count = this.max_count
-    if (this.data.length) msg.since = this.data[this.data.length - 1].time
-
-    this.sock.send(msg)
-    this.subscribed = true
-  }
-
-
-  _unsubscribe() {
-    if (this.sock.connected) this.sock.send(this._get_message('unsubscribe'))
-    this.subscribed = false
-  }
-
-
-  update() {
-    if (!this.subscribed && this.has_subscribers) this._subscribe()
-    if (this.subscribed && !this.has_subscribers) this._unsubscribe()
-  }
-
-
-  add_data(data) {
-    Object.values(this.subscribers).map(cb => {data.map(cb)})
-    this.data.push(...data)
-
-    // Remove any excess data
-    if (this.max_count < this.data.length)
-      this.data.splice(0, this.data.length - this.max_count)
-  }
-
-
-  on_message(msg) {
-    let data = Array.isArray(msg.data) ? msg.data.reverse() : [msg.data]
-    this.add_data(data)
-  }
-
-
-  on_open()  {this.update()}
-  on_close() {this.subscribed = false}
 }
 
 
@@ -115,22 +60,22 @@ class APISock extends Sock {
   constructor(ctx, ...args) {
     super(...args)
     this.ctx    = ctx
-    this.teams  = {}
+    this.subs   = {}
     this.nextID = 1
   }
 
 
-  subscribe_team(team, cb) {
-    if (this.teams[team] == undefined)
-      this.teams[team] = new Team(this, team)
+  subscribe(chart, cb) {
+    let sub = get_chart_subscriber(this, chart)
+    let ref = sub.ref
 
-    return this.teams[team].add_subscriber(cb)
+    if (this.subs[ref] == undefined) this.subs[ref] = sub
+
+    return {ref, id: this.subs[ref].add_subscriber(cb)}
   }
 
 
-  unsubscribe(id) {
-    Object.values(this.teams).map(t => t.del_subscriber(id))
-  }
+  unsubscribe(o) {this.subs[o.ref].del_subscriber(o.id)}
 
 
   on_message(msg) {
@@ -140,19 +85,19 @@ class APISock extends Sock {
       return
     }
 
-    if ((msg.$ref || '').substr(0, 5) == 'team-')
-      this.teams[parseInt(msg.$ref.substr(5))].on_message(msg)
+    let sub = this.subs[msg.$ref]
+    if (sub != undefined) sub.on_message(msg)
 
     else throw 'Unsupported API Websocket message: ' + JSON.stringify(msg)
   }
 
 
-  on_open(event) {Object.values(this.teams).map(t => t.on_open(event))}
+  on_open(event) {Object.values(this.subs).map(t => t.on_open(event))}
 
 
   on_close(event) {
     setTimeout(() => this.connect(), 1000)
-    Object.values(this.teams).map(t => t.on_close(event))
+    Object.values(this.subs).map(t => t.on_close(event))
   }